refactor(contracts): move types + breaking_change_detector + field_tracking from core (#604)
Three of the seven moveable contract_drift modules migrate to
mockforge-contracts. Their inter-dependency forces the group:
- types.rs depends on BreakingChangeDetector
- breaking_change_detector.rs depends on types::DriftBudget etc.
- field_tracking.rs is standalone

Imports rewritten in the moved files to source `Mismatch` /
`ContractDiffResult` from `mockforge_foundation::contract_diff_types`
directly (the existing `crate::ai_contract_diff::*` path resolved
through a re-export chain into foundation anyway).

Internal core caller `budget_engine.rs` repointed to source the
moved types from `mockforge_contracts::contract_drift::*`.

Forwarding re-exports added to core's contract_drift/mod.rs so
that `crate::contract_drift::{DriftBudgetConfig, …}` paths used
by other core modules (config, workspace, voice_workspace) keep
resolving without requiring those call-sites to be touched.

No external callers needed repointing (verified).

The contracts mod.rs gets the new declarations + re-exports + an
updated docstring bullet list. The remaining "NOTE: The following
remain in core" block is preserved; Task 3 updates it after the
protocol-contract cluster also moves.
